+++ packages/tiny-canvas/src/index.js
@@ -8,3 +8,4 @@ export { default as Mark } from './Mark';
 export { default as Link } from './Link';
 export { default as Image } from './Image';
 export { useResetCanvas } from './useResetCanvas';
+export { slugifyCanvasPageName } from './pageRoute';